requirements.txt
click>=8.0.0
anthropic>=0.3.0
openai>=1.0.0
asyncio-mqtt>=0.11.0
aiofiles>=23.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
click>=8.0.0
anthropic>=0.3.0
openai>=1.0.0
aiofiles>=23.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0